requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
numpy==1.26.1
pillow==9.5.0


